aio fixes and allow READs >2G
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@17406 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ mount_slash/mount_slash.h
@@ -200,7 +200,7 @@ struct slashrpc_cservice *
 	 msl_bmap_to_csvc(struct bmapc_memb *, int);
 void	 msl_bmap_reap_init(struct bmapc_memb *, const struct srt_bmapdesc *);
 int	 msl_dio_cb(struct pscrpc_request *, int, struct pscrpc_async_args *);
-int	 msl_io(struct pscfs_req *, struct msl_fhent *, char *, size_t, off_t, enum rw);
+ssize_t	 msl_io(struct pscfs_req *, struct msl_fhent *, char *, size_t, off_t, enum rw);
 int	 msl_read_cb(struct pscrpc_request *, int, struct pscrpc_async_args *);
 void	 msl_reada_rpc_launch(struct bmap_pagecache_entry **, int);
 int	 msl_readahead_cb(struct pscrpc_request *, struct pscrpc_async_args *);
